day 23: tiny example mostly working
The first two iterations of the tiny example work fine, and
then I get array indexing errors, which is expected.

However, I just realized that I don't need to represent the grid
at all: just a set of elf positions is enough. Every necessary query
can be answered by a membership test.

I'll have to figure out how to do basic printing, but that should be
just a matter of tracking the bounding box which I'll need anyway for
the answer to part 1.