day 16: convert basic state actions into methods
Create several methods on State to hide most detailed knowledge
of it's attributes, in particular, the particular attribute
updates needed for the basic actions a State supports.

- new(): create a well formed state, and calculate
  total_flow from the open set.

All of the following methods decrement the ttl, and increment
the total pressure release.

- wait_here(): No motion.

- open_valve(): Open the given valve, which should be the valve
  at the location of `self`, but this is not checked.

- move_to(): move to the given valve, which should be a neighbour
  of the current location.

These changes should make it easier to introduce another searcher
in the maze.